hotfix: progress abstract class 제거
추상클래스를 제거하기 위해 변경.
